+++ wandb/wandb_controller.py
@@ -63,8 +63,9 @@ import six
 from six.moves import urllib
 import wandb
 from wandb import env
-from wandb import wandb_lib, wandb_sdk
+from wandb import wandb_sdk
 from wandb.apis import InternalApi
+from wandb.util import handle_sweep_config_violations
 import yaml
 
 # wandb.sweeps.sweeps will be loaded later to prevent dependency requirements for non sweep users.
@@ -411,7 +412,7 @@ class _WandbController:
                 raise ControllerError("Validate Error: %s" % msg)
         # Create sweep
         sweep_id, warnings = self._api.upsert_sweep(self._create)
-        wandb_lib.sweepwarn.handle_sweep_config_violations(warnings)
+        handle_sweep_config_violations(warnings)
 
         print("Create sweep with ID:", sweep_id)
         sweep_url = _get_sweep_url(self._api, sweep_id)
@@ -469,7 +470,7 @@ class _WandbController:
         _, warnings = self._api.upsert_sweep(
             self._sweep_config, controller=controller, obj_id=sweep_obj_id
         )
-        wandb_lib.sweepwarn.handle_sweep_config_violations(warnings)
+        handle_sweep_config_violations(warnings)
         self._controller_prev_step = self._controller.copy()
 
     def _start_if_not_started(self):
@@ -802,7 +803,7 @@ def sweep(sweep, entity=None, project=None):
     wandb_sdk.wandb_login._login(_silent=True)
     api = InternalApi()
     sweep_id, warnings = api.upsert_sweep(sweep)
-    wandb_lib.sweepwarn.handle_sweep_config_violations(warnings)
+    handle_sweep_config_violations(warnings)
     print("Create sweep with ID:", sweep_id)
     sweep_url = _get_sweep_url(api, sweep_id)
     if sweep_url:
